% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMLMoE.R
\docType{class}
\name{ModelMLMoE-class}
\alias{ModelMLMoE-class}
\alias{ModelMLMoE}
\title{A Reference Class which represents a fitted NMoE model.}
\description{
ModelNMoE represents an estimated NMoE model.
}
\section{Fields}{

\describe{
\item{\code{param}}{A [ParamNMoE][ParamNMoE] object. It contains the estimated
values of the parameters.}

\item{\code{stat}}{A [StatNMoE][StatNMoE] object. It contains all the statistics
associated to the NMoE model.}
}}


\examples{
data(tempanomalies)
x <- tempanomalies$Year
y <- tempanomalies$AnnualAnomaly

nmoe <- emNMoE(X = x, Y = y, K = 2, p = 1, verbose = TRUE)

# nmoe is a ModelNMoE object. It contains some methods such as 'summary' and 'plot'
nmoe$summary()
nmoe$plot()

# nmoe has also two fields, stat and param which are reference classes as well

# Log-likelihood:
nmoe$stat$loglik

# Parameters of the polynomial regressions:
nmoe$param$beta
}
\seealso{
[ParamNMoE], [StatNMoE]
}
