% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamMLMoE.R
\docType{class}
\name{ParamMLMoE-class}
\alias{ParamMLMoE-class}
\alias{ParamMLMoE}
\title{A Reference Class which contains parameters of a NMoE model.}
\description{
ParamMLMoE contains all the parameters of a NMoE model.
}
\section{Fields}{

\describe{
\item{\code{X}}{Numeric vector of length \emph{n} representing the covariates/inputs
\eqn{x_{1},\dots,x_{n}}.}

\item{\code{Y}}{Numeric vector of length \emph{n} representing the observed
response/output \eqn{y_{1},\dots,y_{n}}.}

\item{\code{n}}{Numeric. Length of the response/output vector `Y`.}

\item{\code{K}}{The number of experts.}

\item{\code{p}}{The order of the polynomial regression for the experts.}

\item{\code{q}}{The order of the logistic regression for the gating network.}

\item{\code{wk}}{Parameters of the gating network. \eqn{\boldsymbol{\wk} =
(\boldsymbol{\wk}_{1},\dots,\boldsymbol{\wk}_{K-1})}{\wk =
(\wk_{1},\dots,\wk_{K-1})} is a matrix of dimension \eqn{(q + 1, K -
1)}, with `q` the order of the logistic regression for the gating network.
`q` is fixed to 1 by default.}

\item{\code{eta}}{Polynomial regressions coefficients for each expert.
\eqn{\boldsymbol{\eta} =
(\boldsymbol{\eta}_{1},\dots,\boldsymbol{\eta}_{K})}{\eta =
(\beta_{1},\dots,\beta_{K})} is a matrix of dimension \eqn{(p + 1, K)},
with `p` the order of the polynomial regression. `p` is fixed to 1 by
default.}

\item{\code{df}}{The degree of freedom of the NMoE model representing the complexity
of the model.}
}}

\section{Methods}{

\describe{
\item{\code{initParam(segmental = FALSE, nearTrue, wk_star, etak_star)}}{Method to initialize parameters \code{wk}, \code{eta} and
      \code{sigma2}.

      If \code{segmental = TRUE} then \code{wk}, \code{eta} and
      \code{sigma2} are initialized by clustering the response \code{Y}
      uniformly into \code{K} contiguous segments. Otherwise, \code{wk},
      \code{eta} and \code{sigma2} are initialized by clustering randomly
      the response \code{Y} into \code{K} segments.}
}}

